更新版本至v1.13.3.35版 1.加入Buffer Port Enable/Disable上報的功能 2.修正IMS無法提前趕車的問題(因Nats功能被關閉)

+++ OverhaedHoistTransporter_CSOT/ScriptControl/WebAPI/VehicleInfo.cs
@@ -74,7 +74,7 @@ namespace com.mirle.ibg3k0.sc.WebAPI
                     isSuccess = Enum.TryParse(status, out service_status);
                     if (isSuccess)
                     {
-                        isSuccess = scApp.PortStationService.doUpdatePortStationServiceStatus(port_id, (int)service_status);
+                        isSuccess = scApp.PortStationService.doUpdatePortStationServiceStatus(port_id, service_status);
                     }
                 }
                 catch (Exception ex)
